changed the call to the backend
Now if we press the run button, we can have three scenarios: - run a query -> nothing special, runs a
query, shows the results. it checks for the results every 30 mins. - a query is already running ->
something is already running, blocking another run - this query already ran less than 30 mins ago -> this
query already ran, you can fetch the results of the previous run or try again in a few minutes. The
timeout is calculated, but not a live counter